requirements.txt
# Core OLED display library
luma.oled>=3.12.0

# Image processing for display rendering
Pillow>=10.0.0

# Async HTTP client for API calls
aiohttp>=3.8.0

# Async programming support
asyncio-mqtt>=0.16.0

# Task scheduling
APScheduler>=3.10.0

# HTTP requests (fallback/sync option)
requests>=2.31.0

# GPIO control (Pi-specific)
RPi.GPIO>=0.7.1

# SPI interface support
spidev>=3.6 